requirements.txt
numpy==1.25.2
opencv-python==4.8.0.76
paho-mqtt==1.6.1
